osgi: Support for new OSGi Referenced annotation
Source code can now use @Referenced(SomeClass.class) to make a reference
to the specified types. This will cause Bnd to record a reference to the
type which can result in an Import-Package clause.

The Referenced annotation allows a non-code reference in source code.
This can be helpful if the referenced types are not directly referenced
in the source code but may be needed for some reflective purpose at
runtime and the developer wants to ensure that the necessary
Import-Package clause is generated.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>