central: Better handle multiple projects referring to build.bnd file
The logic for finding an IFile for the workspace build file (e.g.
cnf/build.bnd) failed to address the case were multiple projects could
refer to the file. The main example is where the root of the workspace
is an Eclipse project which can be done to allow easy editing of files
in the root of the repository.

So now we better handle this case and just pick one of the IFile
references.

Fixes https://github.com/bndtools/bnd/issues/4140

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>