[insert] Adds a number of insert menus
- Insert plugin based on the bndplugin annotation
- Syntax based
  - Insert instruction example
  - Insert header
  - Insert macro

This patch uses the Internal Plugin Namespace to find
the name and optional configuration of the plugins. It
then adds the plugin to the existing -plugin or creates
a new one. This is done using BndEditModel so it should
follow all the rules

While being there, I also added insert menus based 
on the Syntax class. 

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>