[launchpad] Test support
Created a constructor where we can pass the RemoteWorkspace.
This makes it a lot easier to setup up small test workspaces.

Not all run specification have a launcher setup, this
was found in adding test cases. The test did not
contain biz.aQute.launcher and failed.

This makes the launcher optional


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>