[resolve mode] Added batch resolve mode
The `before launch` resolve mode will always
resolve before launching. In many scenarios where
a developer is busy, this is an annoying lag.

The `batch` mode will always resolve when in batch
mode or when the -runbundles is empty. This means that
in interactive mode it will not automatically resolve.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>