ds annotations: Better handle osgi.service for factory components
For factory components, we now emit an osgi.service capability for
ComponentFactory with the component.factory attribute set. We no longer
emit an osgi.service capability for any service which can be provided
by an instance of the factory component.

For references to ComponentFactory, we now emit an osgi.service
requirement for ComponentFactory only if the target element of the
Reference annotation is specified. The filter directive of the emitted
osgi.service requirement for ComponentFactory will include the target
filter expression.

Fixes https://github.com/bndtools/bnd/issues/4243

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>