jareditor: Use `EFS.getStore` to access URIs for printing jars
We use a JarFileSystem to view jar files. But the `jarf` scheme is only
known to the Eclipse File System (EFS) and not the JRE. So converting a
`jarf` URI to a URL fails. We use `EFS.getStore` to access URIs for
printing the jar. `EFS.getStore` also properly handles `file` URIs
for non-nested jar files.

Fixes https://github.com/bndtools/bnd/issues/4276

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>