maven: Support bundle for maven must use same versions as maven plugins
The support bundle used versions from the Eclipse m2e for 2018-12.
These versions are newer than that of 3.1.1 which is used as the base
maven version of the maven plugins. So this prevents the maven plugins
from operating on maven version prior to 3.5.3 which is a long way
from 3.1.1. So we change to use the same versions for the support bundle
as for the maven plugins.

Fixes https://github.com/bndtools/bnd/issues/4327

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>