[quickfix] Refactor tests for speed
The test suite for the quick fix processor was quite slow, taking
over a minute to run. On inspection, I noticed that most of the
delay seemed to be in making modifications to the buildpath of
the test project. This fix groups all the tests that have the
same buildpath configuration into the same class, and then modifies
the buildpath in beforeAll() rather than beforeEach(). This brings
total execution time to under 10s rather than over 1m.

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>