[shader] Rename packages
Adding a function to rename packages.

	-shade (<package-spec> ( ';prefix:=' <prefix> )? )+

I've added a rename method to the awesome ClassFile,
and a Shader in the osgi package, although it is not
much code now.

I did not add the code to the ClassFile to keep it clean,
a ClassFileRenamer holds the nitty gritty details. It
contains a SignatureParser that is not creating
objects and is extremely fast, especially when
no renaming takes place. I ran it against 350.000
signatures.

Testing uses a new generator that can generate type
names based on a regex. I think that can be a very
useful tool.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>