Fixed Windows paths not properly
This bug occurs when sourceSets is used in Gradle. This unfortunately cannot be fixed on an exterior basis and requires the following transformation to be made. Since this specific method exclusively calls paths, it should not contain and escaping characters anyways. Henceforth, we can replace such by a proper regex-friendly name.

Signed-off-by: Ghast <shanyujuneja@gmail.com>