[shader] Changes at the request of BJ
BJ hated the visitors so I removed them. Instead
I introduced enums for Info and Attribute so you can 
get a warning if a new type is introduced in the case
statement where they are used.

Moved the ClassFileRenamer to a renamer package
and removed the ClassFile.rename method. Seems
a bit inconvenient but I can live with it.







Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>