jpms: Make sure ModulePackages attribute includes all packages
The wording in
https://docs.oracle.com/javase/specs/jvms/se11/html/jvms-4.html#jvms-4.7.26
is unclear about which packages in a module must be listed in the
ModulePackages attribute. It implies that packages which are not
exported, opened, or service implementation providers may not be
included. But advice in https://bugs.openjdk.java.net/browse/JDK-8255854
states that "The intention is that ModulePackages attribute, if present,
must list all packages in the module".

So we update the JPMS module info generation code to make sure all
contained packages in the bundle are listed in the generated
ModulePackages attribute.

Fixes https://github.com/bndtools/bnd/issues/4398

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>