unmodifiable: Add tests for objects whose hashcode is Integer.MIN_VALUE
If we had uses Math.abs to fold the hashcode value into a non-negative
number for use as an array index, we would be in trouble since
Math.abs(Integer.MIN_VALUE) == Integer.MIN_VALUE. That is, Math.abs can
return a negative number in this case!

So we add some tests to prevent future regression.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>