[remote] Caching of run path
The remote launcher did not properly set -runsystempackages
from the exported packages on the -runpath. This worked
on the first prepare, but the prepare did not clear the
class path. A second prepare cleared -runsystempackages
by recalculating it, but then the not cleared class path
prevented the analysis to see if it was exporting 
anything.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>