maven central check: Improve checking and add tests
We convert the plugin to a VerifierPlugin and enhance the header
checks. The verifier plugin assumes that the proper POM is created from
the manifest headers.

The test verifies warnings are emitted when headers are missing and
also does some checks on the generated POMs.

Fixes https://github.com/bndtools/bnd/issues/1479

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>