[hover] BndEditor macro expansion
The macro expansion was not handling the expansion
of ${.} and the base correctly because the
BndEditModel did not properly set it.

I cleaned up the getProperties() method to be in
line with the use of inheritance of Processors. 
This made one test case fail. 

The BndEdit model has maintains a delta list for
the editor and a set of properties. In the previous
implementation of the getProperties() method the
returned Processor had these properties + the delta
as its properties. However, it also inherited from
the parent processor, which had the originals, which
by definition are supposed to be the same.

I removed this redundancy because I did not like the
shadowing and it made the code simpler. So
the untouched properties are now directly inherited from
the BndEditModel's project or workspace.

This caused a test in BndEditModelTest to fail 
because it assumed that after it saved the changes
to a document the removed properties would contain
the changes. However, the logic is that after a save
the BndEditMode's parent is updated. This normally
happens through a reload.

I changed the test case to do this reloading. This
is a slight incompatibility but I do not think 
this is a problem in the normal usage of the BndEditModel.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>