[build] Workspace per version defaults
This patch provides the defaults for the properties
for a workspace. These defaults are version specific
and the user can set a `-versiondefaults <v>` 
instruction to select a specific version. This
allows a user to upgrade to a new workspace but
keep the old defaults.

The idea is that new features are written to be
disabled by default. Preferably, this should
done with a scope model if application. That is,
like for example `contract *` . These instructions
are _scoped_. By setting an appropriate default
you can enable these instructions.

By default, a workspace will read in the defaults
of the current version. We maintain already 
a current version in About. However, with the
`-versiondefaults` instruction the user can pick
an earlier version of defaults.

The idea is that we continue to add features but
strictly write them disabled. The version defaults
can then selectively enable them so that the majority
of our users get the benefits. 

The version default file will (5.3.0.bnd) will have
the full set of defaults and will therefore encompass
the previous version's defaults.

However, an organization that upgrades their bnd,
will be able to remain at an older version feature
wise.

This was mainly driven because an older workspace
that I had was upgraded. After the upgrade, the 
artifacts failed in the field because the maven
dependencies were automatically added to the pom. Similar
problems happened with the new Java import and I 
think I once messed up with the EE requirements based
on the Java class levels.

The prime advantages I see are:

- A more consistent feature model (disable in code, 
  enable by version defaults)
- Concise documentation of the presence of new features in 
  the version default file. This can guide users to
  override the defaults in their workspace file.
- A user choice to revert back to an older feature
  set while being able to take advantage of bug fixes

This code should work as is in the Gradle workspace
build. The gradle native plugin & maven need to 
see how they get their current defaults. Since the
defaults were already maintained this way, I expect
that this requires no change but I know little about
these uses of bndlib.

- Made versions ignore micro updates
- added test cases & set version defaults for standalone 
and default case

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>