[resource] Main-Class attribute not trimmed
The Main-Class attribute was not trimmed. This
caused extraneous spaces to be at the end sometimes.

This trims the Main-Class and Bundle-Version now.

Also unrolled a stream loop because the check for
capabilities was 20 levels deep in a stream. This makes it impossible to single step or debug.

Unless there  is a demonstrable _and_ significant speed gain we should stop replacing plain old loops by streaming. Not only because it is hard to debug, it also looses the earlier context when you want to report something. Maybe if Java had true closures and no checked exception it might have been useful.





Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>