instruction: Improve thread-safety around reads and writes of pattern field
The pattern() method seemed to be returning null which should not be
possible absent some thread-safety issue. We don't need to synchronize
on the write to the pattern field since writing it more then once is not
harmful. I restructured the implementation to be careful to avoid the
possibility of returning null.

Fixes https://github.com/bndtools/bnd/issues/4461

Signed-off-by: BJ Hargrave <bj@hargrave.dev>