gradle: Add -dependson to test task inputs
The test task was already dependent on the -buildpath and -testpath
through the test task's dependencies on the compileJava and
compileTestJava tasks, respectively. We further ensure that the test
task is also dependent on the jars of any additional projects specified
through -dependson.

So like Eclipse, we now ensure all dependency jars (-buildpath,
-testpath, and -dependson) are built before the test task runs.

Fixes https://github.com/bndtools/bnd/issues/4484

Signed-off-by: BJ Hargrave <bj@hargrave.dev>