[repoviewer] Suppress bundles under GAV pseudo-BSNs
Prior to this fix, if there were bundle and non-bundle versions of a jar
in (eg) a MavenBndRepository, there would be two root entries - one
with the full BSN, and the other with a pseudo-BSN of the form group:artifact.
The plain jars would appear under the pseudo-BSN while the true bundles
would appear under both. With this fix, the bundles will be filtered from the
G:A node so that they only appear once under their actual BSN.

Fixes #3383.

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>