ds: Fix handling of ignored members when processing classes
We ignore certain members when looking for DS annotations such as static
members, synthetic methods, etc. However we process annotations on
ignored members using a stale value in the member field.
So now we avoid stale values in the member field and we check
if the member is ignored before processing its annotations.

Fixes https://github.com/bndtools/bnd/issues/4491

Signed-off-by: BJ Hargrave <bj@hargrave.dev>