[bndtools explorer] Refresh workspace from disk
When the workspace is refreshed from the Bndtools 
explorer, it will do the following:

- Stop autobuild
- Add any projects that are in the bnd workspace
  but not in the Eclipse workspace
- Delete any projects in the Eclipse workspace
  that are located in the bnd workspace but are no
  longer found
- Clean all projects the bnd way 
- Update _all_ the .classpath files
- Refresh the whole workspace
- Set the workspace build order to the bnd 
  build workspace build order
- Set concurrent build jobs to 10
- Do a clean build
- Restore autobuild

This fixes #4413

This patch also fixes  #4434. When compiler errors
are detected, the error is marked directly in the
Eclipse workspace. This used to be reported as an
error on the Project. There was a bug in generate
that did not clear the bnd errors and then this
caused the problem that some fixed errors were still
counted.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>