[resolver] Increase number of ticks in resolver progress
The initial implementation of the resolver progress would allocation 1% of
the remaining ticks for each new requirement. I noticed that on any resolve
of sufficient complexity that the progress bar was actually used, this still
resulted in the progress getting to 100% very quickly even when there was
still a fair bit of work to do. Increasing this to 0.1% on each requirement
make the progress bar a bit more useful.

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>