[builder] Dynamic References
We were using a deprecated API to set our 
dependencies. This caused us to sometimes
postpone the build if we had not provided our
proper dependencies yet.

It turned out there is an option in the builder
extension point to provide a helper that can
calculate the dependencies.

This path sets the plugin and and provides the
helper.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>