deps [bndtools test] Change "synchronous" implementations to run in background
Original implementation of these tests assumed that when an IProgressMonitor was passed in,
the client would always call done(). This is not the case - it's actually the outermost caller'"'"'s
responsibility to call done(). Further, when I wrote the code, I understood (incorrectly)
that these methods that took a monitor were run asynchronously - this is not the case, they are
run in the calling thread.

New implementation spawns a background job, runs the task in the job, and then waits for the
job to finish. This allows the synchronously() implementation to guard against deadlocks
in the sub tasks and report them back up as errors if they take too long.

It is possible that this is overkill and the entire test code could be simplified by calling
the tasks directly with a null progress monitor and abandoning the use of synchronously()
(which it seems never did what I thought it should do anyway). Originally synchronously() was
introduced to attempt to debug lockup issues that I was having which were associated with
issue #4253 - these have now been pinpointed to an earlier part in the code which is
appropriately commented.

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>