gradle: Fix test case to properly configure task dependencies
Gradle 7 warns that we did not have them properly configured. We had
compensated for this mistake by causing the dependencies to be built
first from the command line.

Also, in addition to the warning, Gradle 7 now reruns the tasks which
caused the test to fail since it expected the task to be up-to-date.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>