gradle: Refactor implementation to work with Groovy 2 and 3
When we update the build's gradlew to 7.0, we will be building the
Bnd Gradle plugins with Groovy 3 but we still need to make sure they
run on older Gradle versions which use Groovy 2.

This necessitates some restructuring of the code to avoid some
unnecessary use of closures and to avoid references to private fields
inside the remaining closures. Groovy 2 and 3 seem to handle this
in a different and incompatible way, so we just avoid trying to do it
by copying the private field into a local variable which the closure
can capture.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>