[quickfix test] Manually refresh after updating bnd.bnd
Ran into an issue were a build happened in CI and a bundle wasn't
on the classpath when it should have been. This was a transient problem.
Suspect it may be to do with Eclipse not being notified of the changes
to the bnd.bnd file in a timely manner. Refresh manually in an effort
to fix this.

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>