[nexus] Staging bug in Nexus
Nexus creates a new repository the first time it
sees a release (non-snapshot) to the staging
URL. Subsequent stores in that URL will then go
to this created repo.

The repo depends on the group id and the IP number
of the request. Officially, we're doing this ok.
However, this multiple repos is annoying.

So this patch will upload the _first_ artifact
and then delay 5 secs. 

Ugly, but hopefully working.

I've thought of supporting the staging API but
reverted this code. The API is not very well 
documented and only works on Nexus.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>