[refresh] Cleaned up refresh
There were too many places where we were doing a 
refresh, several inside the bnd lock.

This code moves all refresh of the file system to
one point and that is normally going to be asynchronous
(this is the generate code in the compiler participant).

The refresh will try to coalesce refresh events.

An assert is added to verify that the refresh job is 
never executed in the bnd lock.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>