unmodifiable: Reduce storage use by one third for set and map
We change to use a short array instead of an int array to hold the hash
bucket. This limits the size of the set or map to 2^16-1 entries (since
we use unsigned short values) but that is plenty large.

Tests are added to verify the max size and whether an exception is
thrown if exceeded.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>