bndtools.core: Update bndCall to support after actions
bndCall will supply the operation an object to register an after lock
action. When the operation completes and the bnd lock is released,
bndCall will call each after action in order. This enables work to be
done after the bnd lock is released when that work does not need to be
done while holding the bnd lock. This can avoid deadlocks.

Making this part of bndCall makes it a common feature that callers can
use without rolling their own after lock processing.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>