maven: Separate parent and aggregation poms
The maven folder has a bnd.bnd file and also was the parent pom which
has a .project file naming the project bnd-plugin-parent for m2e. This
violated the convention that bnd projects are named the same as their
folder name.

So we move the parent pom into a subfolder, bnd-plugin-parent, and
change the pom in the maven folder to just be an aggregation pom. We
then also change the .project file to have the maven name.

So this will conform to normal bnd project naming conventions and also
keep the parent pom in an m2e project named bnd-plugin-parent.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>