processor: Refactor plugin support into the PluginsContainer class
The prior logic, in its attempt to avoid deadlocks, could expose a
child processor to zero parent plugins. So this meant that
creating or refreshing a workspace also required the caller to force the
workspace to load its plugins right away. Otherwise, when each
project initialized its plugins, they could attempt to access their
parent's plugins across multiple threads (e.g. gradle parallel build)
and all but one project could see the workspace as having zero plugins.
This has occurred in the OSGi build since it builds bnd plugins in cnf
and then refreshes the workspace to include the just built plugins.

This refactoring addresses this issue at the cost of some breaking API
changes to rarely overridden protected methods due to the
use of the new PluginsContainer type. The new PluginsContainer type
implements both Set<Object> and Registry. Since PluginsContainer
implements Registry, we can pass it to RegistryPlugins at their
customization rather than Processor. This lets us break the cycle
between initializing plugins in the Processor and using the Processor
as a Registry for initializing plugins.

Since we are at a major version boundary, this is a good time to fix
this.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>