repo: Fix BndPomRepository to handle pseudo BSNs like MavenBndRepository
We use the Artifact.getWithoutVersion() value as the pseudo BSN for the
bnd.info capability. We also use Artifact.toString() as the "from"
attribute value in the bnd.info capability which allows us to recreate
the corresponding Artifact object for get calls on the repo. This
properly handles the case where the bundle version and the artifact
version are not identical which is not uncommon. A test case is
extended to confirm this.

Closes #4674

Signed-off-by: BJ Hargrave <bj@hargrave.dev>