[sync] Project creation
Project creation failed because bndtools code created a new Eclipse project but this interacted badly with the synchronizer.

The bndtools project creation needs
to use the synchronization facilities. This makes the whole process a lot simpler and more reliable. Bndtools should create the project in the workspace and the synchronization then automatically creates the Eclipse projects.

However, this is non-trivial since it requires a new dialog for the bndtools project creation (which can then include the template selection).

So this patch disables the auto sync feature until this is done.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>