gradle: Replace use of newly deprecated Gradle items in BndPlugin
Gradle 7.1 has deprecated several methods and the use of conventions.
So we update to use the replacement methods which were introduced in
Gradle 6.1.

We also now handle SourceDirectorySets added as extension to SourceSets.
Prior to Gradle 7.1, these were done as conventions to SourceSets. So
we now handle both ways of adding SourceDirectorySets.

Recent deprecations in Gradle require us to use methods introduced in
Gradle 6.1. So we update the minimum supported Gradle version to
6.1.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>