gradle: Deprecate FileSetRepositoryConvention
As Gradle are moving to deprecate conventions, we need to react and
replace the use of conventions. This change moves
the bundles property into the Bndrun task type.

The existing convention is still in place but unused. It delegates to
the Bndrun task. In the Gradle 8 timeframe, when Gradle removes
convention support, we will need to remove FileSetRepositoryConvention.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>