gradle: Refactor Bnd Gradle Plugins build
We move the Bnd Gradle Plugins into the gradle-plugins folder where
it now has a "normal" gradle build. This build is composited with the
current Bnd build. Artifacts from the current Bnd build are consumed by
the gradle-plugins build via the Maven Local repo.

The new gradle-plugins build enables the ability to use the Eclipse IDE,
via Buildship, for code development of the gradle plugins and
compilation against the Gradle API. This will then support the
converting Groovy classes to Java classes.

It also allows the introduction of Kotlin DSL code for the
BundleTaskExtension used by the Jar and Bundle tasks.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>