gradle: Add Kotlin DSL for bundle extension of Jar and Bundle tasks
This required changing the build script to Kotlin and some hackery so
that we can build the Groovy code before the Kotlin code since the
Kotlin DSL code requires access to the main classes from Groovy.

We also have to use a separate output director for the Kotlin code
since the Kotlin compile task seem to delete the output folder before
it starts!

Signed-off-by: BJ Hargrave <bj@hargrave.dev>