org.bndtools.launch: Fix break in the p2 installation
We disable importing of java.* packages which upsets the p2 packaging
code. This occurred because the bnd.bnd file for the new
org.bndtools.launch project did not put osgi.core on the buildpath.

As a safety measure, we also now use -noimportjava=true.

Also: fixed compile warning in Eclipse for references to code in the
bndtools.utils project, and removed the improper inclusion and exporting
of the bndtools.launch.util package from bndtools.core.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>