[external plugin] Support external plugins in -plugin
Until now, the -plugin only accepted concrete classes. 
This patch allows you to register an interface with
an optional glob for an external plugin name.

After the plugins are initialized, the workspace will
try to find external plugins for the interfaces in the
-plugin. It will link these as proxies. If the proxy
is used, it will load the plugin. 

Closable and Plugin interfaces are supported. For
Plugin, the attributes are used for setProperties.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>