[abstract plugins] Concurrency
Made the closable plugins concurrent safe. Although
the current usage in the workspace guarantees this 
is not a problem, addClosable is a public method 
so it can now be called at any time.

I tried to do the check for closable in add() but
unfortunately some test code assumed the added
objects were not closed. This would prohibit checking
the closeableness at close so we did not need a 
separate set.

Anyway, it is now a copy on write set

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>