bndtools: Don't refresh projects from the repository view refresher
The repository view refresher it to update the view when repositories
change. It itself must not cause changes to the model since we can then
end up in loops which waste lots of time.

Refreshing projects needs to be done outside of the view refresher when
some action updates a repository. Refreshing the view and refreshing
the projects are independent operations which cascade from refreshing
a repository.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>