launcher: Update to avoid output interleaving
We are still seeing some output interleaving in tests.

    15    12  ACTIV 202110062023  Running in main
    reference:file:/home/runner/work/bnd/bnd/biz.aQute.bndall.tests/...

So we update to create each line as a String and then output the String
since that appears to be better synchronized at the innermost stream.

The tests which depend upon the output are also updated also. The output
is split by line before searching for expected results rather than
searching over the complete output. This simplifies the regex since we
don't need to worry about matching across lines.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>