remoteinstall: Clean up build and source code
Eclipse showed a number of compile warning. This is partly due to the
code and partly due to the missing Eclipse settings. The code is
formatted with the settings.

We also make sure org.eclipse.swt is listed in -buildpath before the
swt fragment (which hold the proper code). This makes sure the
import package clauses reference the org.eclipse.swt bundle symbolic
name and not the fragments. We also add the bundle to
bndtools.core.test bndruns to make sure it resolves in the constructed
Eclipse instance.

We replace use of GSON with JSONCodec to avoid an external dependency.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>