remoteinstall: The feature was incomplete and could not install
The remoteinstall bundle depended upon the remote.api bundle which
is not part of the main feature. So p2 cannot install the feature.
The remote.api bundle is not really well formed in that it duplicates
exports from bndlib which is already part of the main feature.

Since the remoteinstall bundle needed very little of the remote.api
code, we now just conditional package it into the remoteinstall bundle
to avoid another dependency.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>