remoteworkspace: Regenerate old port file
Sometimes old port files do not get deleted when the VM exits. So
if a later run picks a free port which is represented by an undeleted,
old port file, we need to update the lastModified time so that the
client will properly sort it as most recent and not use the server
port for an older remote workspace.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>