central: Refresh workspace inside bndLock for cnf change
If a change to cnf was detected, we called workspace refresh without
regard to any running builders. This caused issues where the workspace
refresh closed repos while a builder was running.

We now refresh the workspace while holding the bndLock. We use a job
to handle this in case we need to wait for the lock and don't want to
block the notification (which could be the main) thread.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>