connection: Update ConnectionSettings to match simple names for URLs
We update the settings to allow simple (not-URLs) ids to match URLs.
Maven complains about certain chars used in ids.

`must not contain any of these characters \/:"<>|?*`

To avoid this warning and to be able to share settings.xml files between
maven and Bnd workspace builds, we now match simple ids to URLs.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>