[abstract plugins] lazy strategy
It was too hard to control the creation order from
the outside in the Processor. Changed to a strategy
where abstract plugins are lazily loaded. At
interpretation time of the -plugin instruction,
a special PluginProvider is added. Then when the 
specific type of that provider is wanted, we 
create the implementations.

Probably a bit faster and at least it is now not
mucking with the initialization order.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>