getopt: Declare Target for annotations
This is necessary to address a compiler difference between ecj and
javac. Without the Target, javac would generate a
RuntimeVisibleAnnotation but ecj would generate a
RuntimeVisibleAnnotation and a RuntimeVisibleTypeAnnotation for the
method's return type.

This caused a baselining mismatch in Bndtools when it baselined the
ecj generated classes against the javac generated classes in the
baseline jar. So every Bnd release, we get a baselining error in
Bndtools for all packages which used the getopt annotations.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>