plugincontainer: Cleanup PluginProvider impl
We revert some prior changes which are no longer necessary. We enable
access to PluginProvider (as a plugin type since it does exist in the
set). We use memoize to perform the loading which ensures proper
thread safety on the read side of the external plugins list. We also
handle requests for super types of the external plugin type.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>