include: Fix documentation regarding overwriting properties
The code has, since at least 2010, defaulted to overwriting properties
in included files. That is, included files, by default, overwrite
properties in the including properties. A test case expects this
behavior.

The documentation however says the default is not to overwrite. Since
the code's behavior dates back at least 12 years, it would not be safe
to change the code to match the documentation as it would be a breaking
behavioral change. So we update the documentation to document the
actual behavior of the code.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>